requirements.txt
djangorestframework
django 
djangorestframework-simplejwt 
drf-yasg 
setuptools
psycopg2-binary 
environs 

